🔨📚✏️ Refactor Card trait to class and update Deck class
The Card trait has been refactored to a class with added methods for better functionality, including toString, equals, and hashCode. The description in Deck.scala has been updated to specify that the deck is intended for a Yu-Gi-Oh! game. Necessary instructions regarding these changes have been updated in the README.md file.